requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
opencv-python
scikit-learn
matplotlib